requirements.txt
python-telegram-bot
emoji
mwt